refactor: スキーマのファイル配置を変更
