+++ src/domain/article/schema/articleSchema.ts
